requirements.txt
requests
beautifulsoup4
colorama
dnspython
phonenumbers
Pillow
PyPDF2
python-docx
